dags/requirements.txt
apache-airflow-providers-tableau==2.1.8
apache-airflow-providers-amazon==3.4.0
apache-airflow-providers-postgres==4.1.0
sharepy==1.3.0
pdfrw==0.4
pypdf2==1.27.5
openpyxl==3.0.9
xlrd==2.0.1
json2html==1.3.0